changes include: - abilities have combat point cost - abilities particle effects show properly - movement joystick sensistivity increase

+++ Assets/Scripts/PlayerScripts/AbilityScripts/ActiveAbility.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class ActiveAbility : BaseAbility
 {
-    public int _activeCost;
+    public float _activeCost;
     private BaseAbilityEffect _activeEffect;
 
     void Awake()
